Prompts/requirements.txt
# LangChain Core
langchain
langchain-core

# OpenAI
langchain-openai
openai

# Anthropic
langchain-anthropic

# Google Gemini
langchain-google-genai
google-generativeai

# HuggingFace
langchain-huggingface
transformers
huggingface-hub

# Utils
python-dotenv
numpy
scikit-learn
